+++ changelog.md
@@ -1,3 +1,9 @@
+### 1.2.0
+
+- Added `bundleRelease` and prefix assembleRelease like so `app:assembleRelease`
+- Added build number to apk filename.  
+  Eg: `app-release-${appId}-${appVer}-b${appBuild}.apk`
+
 ### 1.1.0
 
 - Added build number to apk filename.